requirements.txt
streamlit>=1.28.0
requests>=2.31.0
google-genai
pyyaml
dotenv
bedrock-agentcore 
strands-agents
bedrock-agentcore-starter-toolkit
